bgremover/requirements.txt
Django==5.0.4
numpy==1.23.5
Pillow==9.5.0
onnxruntime==1.14.1
rembg[cpu]==2.0.50
django-crispy-forms
crispy-bootstrap4
opencv-python
torch
torchvision
imagehash
filetype
scikit-image
pooch 